#368 Cleaned up file locations & added javadocs
